[CLI] Move page builder into new executable
