+++ src/builder/PagesBuilder.js
@@ -1,13 +1,25 @@
-import {readFileSync} from 'fs';
-import yaml from 'js-yaml';
-
-import loadEntries from './loadEntries';
-
-import renderEntries from './render/renderEntries';
-import renderList from './render/renderList';
+import {join as pathJoin} from 'path';
+import {spawn} from 'child_process';
 
 import Builder from './Builder';
 
+const binPath = pathJoin(__dirname, '../../bin/peridot-build-pages');
+
+function spawnBuildPages() {
+  return new Promise((resolve, reject) => {
+    const proc = spawn(binPath, [], {
+      stdio: ['pipe', 'pipe', 'pipe']
+    });
+
+    proc.on('exit', (code) => {
+      if (code !== 0) {
+        reject(proc.stderr.read());
+      } else {
+        resolve();
+      }
+    });
+  });
+}
 
 export default class PagesBuilder extends Builder {
   constructor() {
@@ -17,31 +29,6 @@ export default class PagesBuilder extends Builder {
   }
 
   async build(/*options*/) {
-    const entriesYaml = readFileSync('_entries.yml', {encoding: 'utf8'});
-
-    const entryData = yaml.safeLoad(entriesYaml);
-
-    const entries = await loadEntries(entryData);
-
-    await renderEntries(entries);
-    await renderList(entries);
-  }
-
-  async handleError(err) {
-    if (err.stack) {
-      // JS errors
-      this.log(err.stack);
-
-    } else if (err.url) {
-      // Failed fetch response
-      // e.g. twitter fetch
-      this.log(`${err.url} - ${err.status} ${err.statusText}`);
-      const body = await err.text();
-      this.log(body);
-
-    } else {
-      // Other error
-      this.log(err);
-    }
+    await spawnBuildPages();
   }
 }
