+++ src/pagesBuilder/injectGlobals.js
