+++ src/pagesBuilder/render/renderEntries.js
