+++ src/server/watch.js
@@ -1,12 +1,7 @@
 import {join as pathJoin} from 'path';
-import {spawn} from 'child_process';
 import chokidar from 'chokidar';
 
 import build from '../builder';
-
-import {createProcessLogger} from '../util/logger';
-
-const binPath = pathJoin(__dirname, '../../bin/peridot');
 
 // hmmmm
 let queuedPath = null;
@@ -28,24 +23,9 @@ async function rebuild() {
 
   console.log(`File "${path}" changed...`);
 
-  // spawn rebuild so it uses new components/
-  // TODO: reject on non-zero code to display an error message warning of some sort
-  const buildPagesPromise = new Promise((resolve/*, reject*/) => {
-    const proc = spawn(binPath, ['build', '--skip-webpack', '--skip-copy'], {
-      stdio: ['pipe', 'pipe', 'pipe']
-    });
-
-    createProcessLogger(proc);
-
-    proc.on('exit', resolve);
-  });
-
-  const buildWebpackPromise = build({
-    skipPages: true,
+  await build({
     skipCopy: true
   });
-
-  await* [buildPagesPromise, buildWebpackPromise];
 
   if (queuedPath) {
     rebuild();
